Refactor seat availability to Long and enhance seat selection logic

+++ core/model/src/main/kotlin/core/model/entity/Categorie.kt
@@ -6,5 +6,5 @@ data class Categorie(
     val description: String = "",
     val price: Double,
     val sector: Sector,
-    val available: Int
+    val available: Long
 )
